add the correct image
